Revert to no Library Layout
This commit has all of the main functionality for the pngme base
program working without library refactoring (each module is defined
in its own file and then brought over with use statements). The CLI
arguments, encoding/decoding/remove/print should all be functional as
well. Documentation for the CLI parsing and the main functionality is
not yet complete.